refactor: update import path for syncDIDVC module
- Replaced import of `syncDIDVCCompletion` with `syncDIDVC` in `batch.ts`
- Reflects file renaming for better clarity and consistency
- No functional changes introduced
